requirements.txt
astropy
reproject
specutils
tqdm
rich
opencv-python
numpy
torch
torchvision
pandas
openpyxl
requests
scipy
astroquery
sparclclient
astro-datalab